requirements.txt
flask==3.1.0
opencv-python==4.11.0.86
deepface==0.0.93
mtcnn==0.1.1
numpy==1.26.4
google-generativeai==0.7.2
langchain-google-genai==2.1.2
langchain==0.3.23
pandas==2.2.3
tensorflow==2.19.0
langchain-core==0.3.51
langchain-text-splitters==0.3.8